requirements.txt
viam-sdk==0.29.3
typing-extensions
